+++ src/main/java/dev/willyelton/crystal_tools/tool/SwordLevelableTool.java
@@ -22,6 +22,8 @@ import net.minecraft.world.level.material.Material;
 public class SwordLevelableTool extends LevelableTool {
     private static final float BASE_ATTACK_DAMAGE = tier.getAttackDamageBonus() + 3;
     private static final float BASE_ATTACK_SPEED = -2.4f;
+    private static final float BASE_ATTACK_KNOCKBACK = 0;
+    private static final float BASE_KNOCKBACK_RESISTANCE = 0;
 
     public SwordLevelableTool() {
         // Need an empty block tag
@@ -35,6 +37,14 @@ public class SwordLevelableTool extends LevelableTool {
 
     public float getAttackSpeed(ItemStack stack) {
         return BASE_ATTACK_SPEED + NBTUtils.getFloatOrAddKey(stack, "attack_speed_bonus");
+    }
+
+    public float getAttackKnockback(ItemStack stack) {
+        return BASE_ATTACK_KNOCKBACK + NBTUtils.getFloatOrAddKey(stack, "knockback");
+    }
+
+    public float getKnockbackResistance(ItemStack stack) {
+        return BASE_KNOCKBACK_RESISTANCE + NBTUtils.getFloatOrAddKey(stack, "knockback_resistance");
     }
 
     public boolean canAttackBlock(BlockState pState, Level pLevel, BlockPos pPos, Player pPlayer) {
@@ -80,6 +90,8 @@ public class SwordLevelableTool extends LevelableTool {
             ImmutableMultimap.Builder<Attribute, AttributeModifier> builder = ImmutableMultimap.builder();
             builder.put(Attributes.ATTACK_DAMAGE, new AttributeModifier(BASE_ATTACK_DAMAGE_UUID, "Weapon modifier", this.getAttackDamage(stack), AttributeModifier.Operation.ADDITION));
             builder.put(Attributes.ATTACK_SPEED, new AttributeModifier(BASE_ATTACK_SPEED_UUID, "Weapon modifier", this.getAttackSpeed(stack), AttributeModifier.Operation.ADDITION));
+            builder.put(Attributes.ATTACK_KNOCKBACK, new AttributeModifier("Weapon modifier", this.getAttackKnockback(stack), AttributeModifier.Operation.ADDITION));
+            builder.put(Attributes.KNOCKBACK_RESISTANCE, new AttributeModifier("Weapon modifier", this.getKnockbackResistance(stack), AttributeModifier.Operation.ADDITION));
             return builder.build();
         } else {
             return super.getAttributeModifiers(slot, stack);