+++ changelog.md
@@ -5,6 +5,9 @@
   - Attack Damage
   - Attack Speed
   - Fire Aspect
+  - Knockback
+  - Knockback Resistance
+    - You get knocked back less when holding the sword
 
 ### Changes
 Updated to Forge 1.18.1-39.0.36